refactor(test): group platform image coverage
