+++ test/automation/e2e/e2e-recommendations.test.ts
@@ -130,7 +130,7 @@ describe("E2E recommendation normalizer", () => {
         "scripts/checks/protected-managed-image-contract.ts",
         "tools/e2e/module-tags.mts",
         ".github/workflows/e2e.yaml",
-        "test/vllm-docker-storage.test.ts",
+        "test/platform/images/vllm-docker-storage.test.ts",
       ]) {
         const destination = path.join(tmp, file);
         fs.mkdirSync(path.dirname(destination), { recursive: true });
