+++ test/e2e/support/e2e-semantic-phase-check.test.ts
@@ -108,7 +108,7 @@ describe("semantic E2E phase checker", () => {
         testMatrix: [
           {
             id: "vllm-docker-storage",
-            file: "test/vllm-docker-storage.test.ts",
+            file: "test/platform/images/vllm-docker-storage.test.ts",
             project: "integration",
           },
         ],
@@ -116,7 +116,7 @@ describe("semantic E2E phase checker", () => {
       {
         liveTestToJobs: new Map([
           ["test/e2e/live/bootstrap-install-smoke.test.ts", ["bootstrap-install-smoke"]],
-          ["test/vllm-docker-storage.test.ts", ["vllm-docker-storage"]],
+          ["test/platform/images/vllm-docker-storage.test.ts", ["vllm-docker-storage"]],
         ]),
       },
       [
@@ -133,7 +133,7 @@ describe("semantic E2E phase checker", () => {
       { file: "test/e2e/live/launchable-smoke.test.ts", project: "e2e-live" },
       { file: "test/e2e/live/registry-targets.test.ts", project: "e2e-live" },
       { file: "test/e2e/live/unselected-regression.test.ts", project: "e2e-live" },
-      { file: "test/vllm-docker-storage.test.ts", project: "integration" },
+      { file: "test/platform/images/vllm-docker-storage.test.ts", project: "integration" },
     ]);
   });
