+++ test/platform/images/docker-abstraction-guard.test.ts
@@ -6,7 +6,7 @@ import path from "node:path";
 
 import { describe, expect, it } from "vitest";
 
-const REPO_ROOT = path.join(import.meta.dirname, "..");
+const REPO_ROOT = path.join(import.meta.dirname, "../../..");
 const TSX = path.join(REPO_ROOT, "node_modules", ".bin", "tsx");
 const INVENTORY_SCRIPT = path.join(REPO_ROOT, "scripts", "list-command-helper-uses.mts");
 const DOCKER_ABSTRACTION_PREFIX = "src/lib/adapters/docker/";
