+++ test/platform/images/retry-docker-imagetools-inspect.test.ts
@@ -8,7 +8,7 @@ import path from "node:path";
 
 import { describe, expect, it } from "vitest";
 
-const repoRoot = path.resolve(import.meta.dirname, "..");
+const repoRoot = path.resolve(import.meta.dirname, "../../..");
 const inspector = path.join(repoRoot, "scripts/checks/retry-docker-imagetools-inspect.sh");
 const reference = `ghcr.io/nvidia/nemoclaw/hermes-sandbox-base@sha256:${"a".repeat(64)}`;
